requirements.txt
pystray==0.19.5
PyAutoGUI==0.9.54
keyboard==0.13.5
sympy==1.13.3
pillow==11.1.0
PyYAML==6.0.2
pywin32==310
pyinstaller==6.12.0